feat(ui): move marketplace tab to section and remove block explorer
Marketplace tab is no longer under Smart Contracts section, but in its own section. Block explorer
section is removed.

re #452
