+++ app/renderer/ui/components/main/sections/marketplace/style.scss
@@ -1,6 +1,13 @@
 :local {
   @import "app/renderer/ui/theme";
 
+  .layout {
+    display: grid;
+    grid-gap: 2em;
+    grid-template-columns: 4fr 1fr;
+    padding: 3em;
+  }
+
   .header {
     align-items: center;
     background: $tab_marketplace-header-background;
