+++ app/renderer/ui/components/main/sections/smartContracts/tabs/index.tsx
@@ -1,11 +1,9 @@
 import TabEasyComposer from "./easyComposer"
-import TabMarketplace from "./marketplace"
 import TabMyContracts from "./myContracts"
 import TabProEditor from "./proEditor"
 
 export {
   TabEasyComposer,
-  TabMarketplace,
   TabMyContracts,
   TabProEditor
 }
